+++ lib/database_consistency.rb
@@ -12,7 +12,7 @@ require 'database_consistency/writers/simple_writer'
 require 'database_consistency/checkers/base_checker'
 require 'database_consistency/checkers/table_checker'
 require 'database_consistency/checkers/validator_checker'
-require 'database_consistency/checkers/presence_validation_checker'
+require 'database_consistency/checkers/column_presence_checker'
 require 'database_consistency/checkers/null_constraint_checker'
 
 require 'database_consistency/processors/base_processor'