+++ example/db/migrate/20181028133756_create_countries.rb
@@ -1,0 +1,8 @@
+class CreateCountries < ActiveRecord::Migration[5.2]
+  def change
+    create_table :countries do |t|
+      t.string :name
+      t.timestamps
+    end
+  end
+end