+++ lib/database_consistency.rb
@@ -14,6 +14,7 @@ require 'database_consistency/checkers/table_checker'
 require 'database_consistency/checkers/validator_checker'
 require 'database_consistency/checkers/column_presence_checker'
 require 'database_consistency/checkers/null_constraint_checker'
+require 'database_consistency/checkers/belongs_to_presence_checker'
 
 require 'database_consistency/processors/base_processor'
 require 'database_consistency/processors/models_processor'