+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-### [Master]
+### [0.4.0] - 2019/01/10
+
+Features:
+- Introduce `BelongsToPresenceChecker` to find missed foreign keys.
 
 Breaking changes:
 - Rename `PresenceValidationChecker` to `ColumnPresenceChecker` for simplicity. Please update your configuration files properly.